Added UpdateSuumoBukkenUrlJob
Suumo page for a property (bukken) display listings for that property. For each real estate company there is a separate listing page (chintai/bc_*). These 'bc' pages map many-to-few for overall listing pages (chintai/jnc_*). This job updates the href of these links on the bukken page to point to the 'jnc' listing pages. It also removes duplicates on the bukken page.