Various architectural changes to extension
Jobs now run as a dynamic pipeline of promises. This ensures each job runs one after the other synchronously. This is important when one job depends on the output of another. The scrapedElems object from the loaders is passed from job to job. In addition, a utils module was added for the Jobs to reduce duplicate code