Added AtHome loader for Extension
This line in HighlightSimilarListings is very important:
77: Object.entries(out).some(([kO, vO]) => kL === kO && String(vO).includes(vL))
The reason I use include is because AtHome address doesn't contain the prefecture. So instead I check that the AtHome address is a substring of the DB address. This change shouldn't affect any other param comparisons, because if they are equal, include is still true